+++ docs/CHANGELOG.rst
@@ -14,11 +14,20 @@ Fixed
 -----
 - Do not raise exception when terminating ``runlistener`` management command
 
- 
 Changed
 -------
 - Make S3 connectors use system credentials when they are not explicitely
   given in settings
+
+
+===================
+28.0.4 - 2021-05-04
+===================
+  
+Fixed
+-----
+- Use per process storage overrides
+
 
 ===================
 28.0.3 - 2021-05-04